Use doubles, instead of floats

+++ MandelbrotSet/fs.frag
@@ -1,14 +1,15 @@
 #version 330
+#extension GL_ARB_gpu_shader_fp64 : enable
 out vec3 color;
 
 uniform int WINDOW_SIZE_X;
 uniform int WINDOW_SIZE_Y;
-uniform float zoomfactor;
-uniform float centerx;
-uniform float centery;
+uniform double zoomfactor;
+uniform double centerx;
+uniform double centery;
 uniform int useriterations;
-float stepx = 4.0 / WINDOW_SIZE_X;
-float stepy = 4.0 / WINDOW_SIZE_Y;
+double stepx = 4.0 / WINDOW_SIZE_X;
+double stepy = 4.0 / WINDOW_SIZE_Y;
 const int MaxIterations = 500;
 
 
@@ -16,18 +17,18 @@ const int MaxIterations = 500;
 
 void main()
 {
-	vec2 z, c;
+	dvec2 z, c;
 	z.x = 0;
 	z.y = 0;
 
-	c.x = (gl_FragCoord.x * stepx-2)*zoomfactor+centerx;
-	c.y = (gl_FragCoord.y * stepy-2)*zoomfactor+centery;
+	c.x = (gl_FragCoord.x * stepx-2.0)*zoomfactor+centerx;
+	c.y = (gl_FragCoord.y * stepy-2.0)*zoomfactor+centery;
 
 	int i;
 	for(i = 0; i < useriterations; i++)
 	{
-		float x = z.x*z.x - z.y*z.y + c.x;
-		float y = 2*(z.x*z.y) + c.y;
+		double x = z.x*z.x - z.y*z.y + c.x;
+		double y = 2*(z.x*z.y) + c.y;
 		z.x = x;
 		z.y = y;
